implements contract attack and flee functions

+++ contracts/obstacles/src/constants.cairo
@@ -26,9 +26,9 @@ mod ObstacleId {
     const HiddenArrow: u8 = 15;
 
     // If you add obstacle, make sure to update MAX_ID below
-    // making this u64 as it's commonly used to select a random obstacle based
-    // on entropy variables which are u64 based
-    const MAX_ID: u64 = 15;
+    // making this u128 as it's commonly used to select a random obstacle based
+    // on entropy variables which are u128 based
+    const MAX_ID: u128 = 15;
 }
 
 mod Slot {
